fix(me-setting): rename VistorSettings to AnonymousSettings
