+++ src/views/Me/Settings/Settings/index.tsx
@@ -10,7 +10,7 @@ import {
 
 // import { redirectToLogin } from '~/common/utils'
 
-import ViewerSetting from '../ViewerSettings'
+import AnonymousSettings from '../AnonymousSettings'
 import AccountSettings from './Account'
 import UISettings from './UI'
 import WalletSettings from './Wallet'
@@ -19,7 +19,7 @@ const Settings = () => {
   const viewer = useContext(ViewerContext)
 
   if (viewer.privateFetched && !viewer.isAuthed) {
-    return <ViewerSetting />
+    return <AnonymousSettings />
   }
 
   return (
